// api/generated/model/recipeDifficulty.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Cheffy API
 * OpenAPI spec version: 1.0.0
 */

export type RecipeDifficulty = typeof RecipeDifficulty[keyof typeof RecipeDifficulty];


export const RecipeDifficulty = {
  EASY: 'EASY',
  MEDIUM: 'MEDIUM',
  HARD: 'HARD',
  EXPERT: 'EXPERT',
} as const;