// api/generated/model/recipeSectionIngredientUnit.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Cheffy API
 * OpenAPI spec version: 1.0.0
 */

export type RecipeSectionIngredientUnit = typeof RecipeSectionIngredientUnit[keyof typeof RecipeSectionIngredientUnit];


export const RecipeSectionIngredientUnit = {
  G: 'G',
  KG: 'KG',
  MG: 'MG',
  OZ: 'OZ',
  LB: 'LB',
  ML: 'ML',
  L: 'L',
  CUP: 'CUP',
  TBSP: 'TBSP',
  TSP: 'TSP',
  UNIT: 'UNIT',
  DOZEN: 'DOZEN',
  PINCH: 'PINCH',
  DRIZZLE: 'DRIZZLE',
  CUBE: 'CUBE',
  PACKAGE: 'PACKAGE',
  CAN: 'CAN',
  BOTTLE: 'BOTTLE',
  BOX: 'BOX',
  TO_TASTE: 'TO_TASTE',
  CLOVE: 'CLOVE',
  BUNCH: 'BUNCH',
  GLASS: 'GLASS',
  LEAF: 'LEAF',
  POT: 'POT',
} as const;