// api/generated/model/recipeYieldUnit.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Cheffy API
 * OpenAPI spec version: 1.0.0
 */

export type RecipeYieldUnit = typeof RecipeYieldUnit[keyof typeof RecipeYieldUnit];


export const RecipeYieldUnit = {
  PORTIONS: 'PORTIONS',
  PEOPLE: 'PEOPLE',
  UNITS: 'UNITS',
  SLICES: 'SLICES',
  PIECES: 'PIECES',
  CUPS: 'CUPS',
  GLASSES: 'GLASSES',
  PLATES: 'PLATES',
  TO_TASTE: 'TO_TASTE',
} as const;